Crate governance_os_primitives

Low level primitives for the governance OS runtime and node.

Enums

CurrencyId
How we represent currencies in the runtime.

Role
The different roles supported by the runtime.

Type Definitions

AccountId
Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance
Balance of an account.

Block
Block type.

BlockNumber
An index to a block.

Hash
A hash of some data used by the chain.

Header
Header type.

Index
Index of a transaction in the chain.

Moment
Type used for expressing timestamp.

Signature
Alias to 512-bit hash when used in the context of a transaction signature on the chain.